wrap waterkit and gist in one function
