+++ waterkit/__init__.py
@@ -6,14 +6,31 @@
 
 from .autodock_map import Map
 from .autogrid import AutoGrid
+from .autogrid import calc_spherical_water_map
 from .forcefield import AutoDockForceField
 from .molecule import Molecule
+from .receptor_prep import PrepareReceptor
 from .spherical_model_map import SphericalWaterMap
 from .sampling import WaterSampler
+from .trajectory_utils import make_trajectory
+from .trajectory_utils import WaterMinimizer
 from .water import Water
 from .water_box import WaterBox
 from .waterkit import WaterKit
+from .wrap_waterkit_and_gist import run_waterkit_and_gist
 
-__all__ = ["Map", "AutoGrid", "AutoDockForceField", "Molecule",
-           "SphericalWaterMap", "WaterSampler", "Water",
-           "WaterBox", "WaterKit"]
+__all__ = [
+    "Map",
+    "AutoGrid",
+    "calc_spherical_water_map",
+    "AutoDockForceField",
+    "Molecule",
+    "PrepareReceptor",
+    "SphericalWaterMap",
+    "WaterSampler",
+    "Water",
+    "WaterBox",
+    "WaterMinimizer",
+    "WaterKit",
+    "run_waterkit_and_gist",
+]
